MacOS 平台 build image编译开发支持

+++ build_img.sh
@@ -1,34 +1,82 @@
 #!/bin/sh
-
+################################################################
+# 	MacOS  : brew install dosfstools
+# 	Ubuntu : apt-get install dosfstools
+#	Usage:
+# 		build_img.sh -m [arch] -fs [ext4|fat32] -file [testcast]
+################################################################
 # default setting
 arch=x86_64
 fs=fat32
 FILE=
 
-if [ -n "$1" ]; then
-	if [ "$1" = "riscv64" ] || [ "$1" = "aarch64" ] || [ "$1" = "x86_64" ]; then # $1 is arch
-		arch=$1
-		if [ -n "$2" ]; then
-			if [ "$2" = "ext4" ] || [ "$2" = "fat32" ]; then # $2 is type of file system 
-				fs=$2
-				if [ -n "$3" ]; then
-					FILE=$3
-				fi
-			else  # $2 is folder in testcases
-				FILE=$2
-			fi
-		fi
-	elif [ "$1" = "ext4" ] || [ "$1" = "fat32" ]; then # $1 is type of file system 
-		fs=$1
-		if [ -n "$2" ]; then
-			FILE=$2
-		fi
-	else # $1 is folder in testcases
-		FILE=$1
-	fi
-fi
+display_help()
+{
+	echo ""
+	echo "./build_img.sh -m [arch] -fs [filesystem] -file [testcast]"
+	echo ""
+	exit 1
+}
 
-if [ ! -n $FILE ] || [ "$FILE" = "" ]; then # use default testcases
+while [ -n "$1" ]; do
+	case $1 in
+		-m)
+			shift
+			arch="$1"
+			;;
+		-fs)
+			shift
+			fs="$1"
+			;;
+		-file)
+			shift
+			FILE="$1"
+			;;
+		riscv64)
+			arch=riscv64
+			;;
+		x86_64)
+			arch=x86_64
+			;;
+		aarch64)
+			arch=aarch64
+			;;
+		fat32)
+			fs=fat32
+			;;
+		ext4)
+			fs=ext4
+			;;
+		sdcard)
+			FILE=sdcard
+			;;
+		gcc)
+			FILE=gcc
+			;;
+		redis)
+			FILE=redis
+			;;
+		testsuits-x86_64-linux-musl)
+			FILE=testsuits-x86_64-linux-musl
+			;;
+		ZLM)
+			FILE=ZLM
+			;;
+		libc-dynamic)
+			FILE=libc-dynamic
+			;;
+		libc-static)
+			FILE=libc-static
+			;;
+		*)
+			display_help
+			;;
+	esac
+	shift
+done
+
+
+if [ -z "$FILE" ]; then # use default testcases
 	if [ "$arch" = "riscv64" ]; then
 		FILE=sdcard
 	elif [ "$arch" = "x86_64" ]; then
@@ -47,7 +95,7 @@ if [ "$FILE" = "testsuits-x86_64-linux-musl" ] && [ ! -e testcases/$FILE ]; then
 	rm $FILE.tgz
 fi
 
-rm disk.img
+rm -f disk.img
 dd if=/dev/zero of=disk.img bs=4M count=30
 
 if [ "$fs" = "ext4" ]; then
@@ -58,11 +106,22 @@ else
 fi
 
 mkdir -p mnt
-sudo mount disk.img mnt
 
-# 根据命令行参数生成对应的测例
-echo "Copying $arch $fs $FILE/* to disk"
-sudo cp -r ./testcases/$FILE/* ./mnt/
-sudo umount mnt
-sudo rm -rf mnt
-sudo chmod 777 disk.img
+#### 添加 MacOS 支持
+os=`uname -s`
+if [ "x$os" = "xDarwin" ];then
+	hdiutil detach mnt > /dev/null 2>&1
+	hdiutil attach disk.img -mountpoint mnt
+	echo "Copying $arch $fs $FILE/* to disk"
+	cp -r ./testcases/$FILE/* ./mnt/
+	hdiutil detach mnt
+	chmod 777 disk.img
+else
+	sudo mount disk.img mnt
+	# 根据命令行参数生成对应的测例
+	echo "Copying $arch $fs $FILE/* to disk"
+	sudo cp -r ./testcases/$FILE/* ./mnt/
+	sudo umount mnt
+	sudo rm -rf mnt
+	sudo chmod 777 disk.img
+fi
